requirements.txt
#Ensure to install below reqs...

scikit-learn - Version: 1.1.3
confluent-kafka - Version: 1.9.2
slackclient - Version: 2.9.3
python-dotenv - Version: 0.17.1